test: added memory & streaming tests
