+++ requirements.txt
@@ -1,35 +1,25 @@
-aiohttp==3.9.5
-aiosignal==1.3.1
-annotated-types==0.6.0
-anyio==4.3.0
-asgiref==3.8.1
-async-timeout==4.0.3
-attrs==23.2.0
-backoff==2.2.1
-backports.tarfile==1.1.1
-bcrypt==4.1.3
-build==1.2.1
-cachetools==5.3.3
-certifi==2024.2.2
-cfgv==3.4.0
-charset-normalizer==3.3.2
-chroma-hnswlib==0.7.3
-chromadb==0.5.0
+google_api_python_client==2.123.0
+googlemaps==4.10.0
+langchain_community==0.0.33
+langchain_core==0.1.43
+numpy==1.26.4
+openai==1.20.0
+pandas==2.2.2
+protobuf==4.25.3
+pydantic==2.7.0
+Pympler==1.0.1
+pytest==8.1.1
+Requests==2.31.0
+torch==2.0.1
+transformers==4.38.1
+wolframalpha==5.0.0
+pre-commit==3.7.0
 click==8.1.7
-coloredlogs==15.0.1
-dataclasses-json==0.6.6
-Deprecated==1.2.14
-distlib==0.3.8
-distro==1.9.0
-dnspython==2.6.1
-email_validator==2.1.1
-exceptiongroup==1.2.1
-fastapi==0.111.0
-fastapi-cli==0.0.3
-filelock==3.14.0
-flatbuffers==24.3.25
-frozenlist==1.4.1
-fsspec==2024.3.1
+python-dotenv==1.0.0
+ollama==0.1.7
+wikipedia==1.4.0
+chromadb==0.5.0
+chroma-hnswlib==0.7.3
 google-ai-generativelanguage==0.6.3
 google-api-core==2.19.0
 google-api-python-client==2.123.0
@@ -41,100 +31,6 @@ googlemaps==4.10.0
 greenlet==3.0.3
 grpcio==1.63.0
 grpcio-status==1.62.2
-h11==0.14.0
-httpcore==1.0.5
-httplib2==0.22.0
-httptools==0.6.1
-httpx==0.25.2
-huggingface-hub==0.23.0
-humanfriendly==10.0
-identify==2.5.36
-idna==3.7
-importlib-metadata==7.0.0
-importlib_resources==6.4.0
-iniconfig==2.0.0
-jaraco.context==5.3.0
-Jinja2==3.1.4
-jsonpatch==1.33
-jsonpointer==2.4
-kubernetes==29.0.0
-langchain-community==0.0.33
-langchain-core==0.1.43
-langsmith==0.1.57
-markdown-it-py==3.0.0
-MarkupSafe==2.1.5
-marshmallow==3.21.2
-mdurl==0.1.2
-mmh3==4.1.0
-monotonic==1.6
-more-itertools==10.2.0
-mpmath==1.3.0
-multidict==6.0.5
-mypy-extensions==1.0.0
-networkx==3.3
-nodeenv==1.8.0
-numpy==1.26.4
-oauthlib==3.2.2
-ollama==0.1.7
-onnxruntime==1.17.3
-openai==1.20.0
-orjson==3.10.3
-overrides==7.7.0
-packaging==23.2
-pandas==2.2.2
-platformdirs==4.2.1
-pluggy==1.5.0
-posthog==3.5.0
-pre-commit==3.7.0
-proto-plus==1.23.0
-protobuf==4.25.3
-pyasn1==0.6.0
-pyasn1_modules==0.4.0
-pydantic==2.7.0
-pydantic_core==2.18.1
-Pygments==2.18.0
-Pympler==1.0.1
-pyparsing==3.1.2
-PyPika==0.48.9
-pyproject_hooks==1.1.0
-pytest==8.1.1
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.0
-python-multipart==0.0.9
-pytz==2024.1
-PyYAML==6.0.1
-regex==2024.5.10
-requests==2.31.0
-requests-oauthlib==2.0.0
-rich==13.7.1
-rsa==4.9
-safetensors==0.4.3
-shellingham==1.5.4
-six==1.16.0
-sniffio==1.3.1
-SQLAlchemy==2.0.30
-starlette==0.37.2
-sympy==1.12
-tenacity==8.3.0
 tiktoken==0.7.0
 tokenizers==0.15.2
-tomli==2.0.1
-torch==2.0.1
-tqdm==4.66.4
-transformers==4.38.1
-typer==0.12.3
-typing-inspect==0.9.0
-typing_extensions==4.11.0
-tzdata==2024.1
-ujson==5.10.0
-uritemplate==4.1.1
-urllib3==2.2.1
-virtualenv==20.26.2
-watchfiles==0.21.0
-websocket-client==1.8.0
-websockets==12.0
-wolframalpha==5.0.0
-wrapt==1.16.0
-xmltodict==0.13.0
-yarl==1.9.4
-zipp==3.18.1
+sentence_transformers==2.7.0
